Error fixed in Post component

+++ Frontend/src/pages/Post.tsx
@@ -4,13 +4,15 @@ import appwriteService from "../appwrite/config";
 import { Button, Container } from "../components";
 import parse from "html-react-parser";
 import { useSelector } from "react-redux";
+import { RootState } from "../store/authSlice";
+import { Models } from "appwrite";
 
 export default function Post() {
-  const [post, setPost] = useState(null);
+  const [post, setPost] = useState<Models.Document>();
   const { slug } = useParams();
   const navigate = useNavigate();
 
-  const userData = useSelector((state) => state.auth.userData);
+  const userData = useSelector((state: RootState) => state.auth.userData);
 
   const isAuthor = post && userData ? post.userId === userData.$id : false;
 
@@ -24,12 +26,14 @@ export default function Post() {
   }, [slug, navigate]);
 
   const deletePost = () => {
-    appwriteService.deletePost(post.$id).then((status) => {
-      if (status) {
-        appwriteService.deleteFile(post.featuredImage);
-        navigate("/");
-      }
-    });
+    if (post) {
+      appwriteService.deletePost(post.$id).then((status) => {
+        if (status) {
+          appwriteService.deleteFile(post.featuredImage);
+          navigate("/");
+        }
+      });
+    }
   };
 
   return post ? (
@@ -37,7 +41,7 @@ export default function Post() {
       <Container>
         <div className="w-full flex justify-center mb-4 relative border rounded-xl p-2">
           <img
-            src={appwriteService.getFilePreview(post.featuredImage)}
+            src={appwriteService.getFilePreview(post.featuredImage).toString()}
             alt={post.title}
             className="rounded-xl"
           />
